requirements.txt
discord.py
django
django-heroku
gunicorn
requests
slackclient
